% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_centrifuge.R
\name{add_names_col}
\alias{add_names_col}
\title{Add taxonomic names to centrifuge table}
\usage{
add_names_col(cent_out, cent_report)
}
\arguments{
\item{cent_out}{Centrifuge table}

\item{cent_report}{Centrifuge report table}
}
\value{
Centrifuge table with names column
}
\description{
Add taxonomic names to centrifuge table
}
