% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble_fasta.R
\name{scramble_fasta}
\alias{scramble_fasta}
\title{Scramble a fasta file}
\usage{
scramble_fasta(fasta)
}
\arguments{
\item{fasta}{Input fasta file}
}
\value{
A scrambled DNAstring object
}
\description{
For each sequence in the file, randomly scramble the sequence
}
