% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_centrifuge.R
\name{tax_id_to_name}
\alias{tax_id_to_name}
\title{Convert taxonomic ID to taxonomic name}
\usage{
tax_id_to_name(tax_id, cent_report)
}
\arguments{
\item{tax_id}{Centrifuge taxonomic ID}

\item{cent_report}{Centrifuge report output}
}
\value{
Taxonomic name
}
\description{
Convert taxonomic ID to taxonomic name
}
