% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble_assembly.R
\name{scramble_fasta}
\alias{scramble_fasta}
\title{Scramble a fasta file}
\usage{
scramble_fasta(fasta, outfile)
}
\arguments{
\item{fasta}{Input fasta file}

\item{outfile}{Output file path}
}
\description{
For each sequence in the file, randomly scramble the sequence and write to
file.
}
